MuseumBot/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
Cython==0.29.14
DateTime==4.3
dialogflow==0.6.0
firebase==3.0.1
firebase-admin==2.17.0
Flask==1.0.3
future==0.17.1
ffmpeg-python==0.2.0
gcloud==0.18.3
google-api-core==1.11.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.1
google-cloud-firestore==1.2.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
googlemaps==3.0.2
grpcio==1.21.1
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
liffpy==0.0.1a3
line-bot-sdk==1.10.0
lxml==4.3.4
MarkupSafe==1.1.1
msgpack==0.6.1
numpy==1.16.4
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycrypto==2.6.1
pydub==0.23.1
python-jwt==3.2.5
pytz==2019.1
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
SoundFile==0.10.2
soupsieve==1.9.1
SpeechRecognition==3.8.1
sseclient==0.0.24
uritemplate==3.0.0
urllib3==1.25.3
wavio==0.0.4
Werkzeug==0.15.4
zope.interface==4.6.0
